fix misc equipment creation/update when a default item exists + documentation

+++ fittrackee/equipments/equipments.py
@@ -12,6 +12,7 @@ from fittrackee.responses import (
     ForbiddenErrorResponse,
     HttpResponse,
     InvalidPayloadErrorResponse,
+    MiscEquipmentInvalidPayloadErrorResponse,
     handle_error_and_return_response,
 )
 from fittrackee.users.models import (
@@ -23,9 +24,12 @@ from fittrackee.utils import decode_short_id
 from fittrackee.visibility_levels import VisibilityLevel
 from fittrackee.workouts.models import Sport, Workout
 
-from .exceptions import InvalidEquipmentsException
+from .exceptions import (
+    InvalidEquipmentsException,
+    MiscEquipmentLimitExceededException,
+)
 from .models import Equipment, EquipmentType, WorkoutEquipment
-from .utils import SPORT_EQUIPMENT_TYPES
+from .utils import MAX_MISC_LIMIT, SPORT_EQUIPMENT_TYPES
 
 equipments_blueprint = Blueprint("equipments", __name__)
 
@@ -78,15 +82,43 @@ def update_user_sport_preferences_if_exist(
     default_for_sport_ids: list[int],
 ) -> None:
     if user_sport_preferences:
-        # remove existing pieces of equipment with the same equipment type
-        db.session.query(UserSportPreferenceEquipment).filter(
-            UserSportPreferenceEquipment.c.user_id == auth_user.id,
-            UserSportPreferenceEquipment.c.sport_id.in_(default_for_sport_ids),
-            (
-                UserSportPreferenceEquipment.c.equipment_type_id
-                == equipment.equipment_type_id
-            ),
-        ).delete()
+        if equipment.equipment_type.label == "Misc":
+            # can not exceeds max limit for Misc type
+            sport_ids = []
+            for sport_id in default_for_sport_ids:
+                query = db.session.query(UserSportPreferenceEquipment).filter(
+                    UserSportPreferenceEquipment.c.user_id == auth_user.id,
+                    UserSportPreferenceEquipment.c.sport_id == sport_id,
+                    (
+                        UserSportPreferenceEquipment.c.equipment_id
+                        != equipment.id
+                    ),
+                    (
+                        UserSportPreferenceEquipment.c.equipment_type_id
+                        == equipment.equipment_type_id
+                    ),
+                )
+                if query.count() >= MAX_MISC_LIMIT:
+                    sport_ids.append(sport_id)
+
+            if sport_ids:
+                raise MiscEquipmentLimitExceededException(
+                    message=f"a maximum of {MAX_MISC_LIMIT} pieces of Misc "
+                    f"equipment can be added",
+                    sport_ids=sport_ids,
+                )
+        else:
+            # remove existing pieces of equipment with the same equipment type
+            db.session.query(UserSportPreferenceEquipment).filter(
+                UserSportPreferenceEquipment.c.user_id == auth_user.id,
+                UserSportPreferenceEquipment.c.sport_id.in_(
+                    default_for_sport_ids
+                ),
+                (
+                    UserSportPreferenceEquipment.c.equipment_type_id
+                    == equipment.equipment_type_id
+                ),
+            ).delete()
 
         db.session.execute(
             insert(UserSportPreferenceEquipment)
@@ -328,6 +360,8 @@ def post_equipment(auth_user: User) -> Union[Tuple[Dict, int], HttpResponse]:
 
     **Example response**:
 
+    - success
+
     .. sourcecode:: http
 
       HTTP/1.1 201 CREATED
@@ -358,6 +392,21 @@ def post_equipment(auth_user: User) -> Union[Tuple[Dict, int], HttpResponse]:
         "status": "created"
       }
 
+    - 'misc' equipment exceeding the max limit, when other equipment items
+      are already associated with sports
+
+    .. sourcecode:: http
+
+      HTTP/1.1 400 BAD REQUEST
+      Content-Type: application/json
+
+      {
+        "message": "a maximum of 5 pieces of Misc equipment can be added",
+        "sport_ids": [1],
+        "status": "limit_exceeded"
+      }
+
+
     :<json string label: a brief (less than 50 characters) label for
         the piece of equipment
     :<json integer equipment_type: the ID for an equipment type (it must be
@@ -367,7 +416,11 @@ def post_equipment(auth_user: User) -> Union[Tuple[Dict, int], HttpResponse]:
     :<json boolean is_active: whether or not this equipment is currently
         active (default: ``true``)
     :<json array of integers default_for_sport_ids: the default sport ids
-        to use for this equipment, not mandatory
+        to use for this equipment, not mandatory.
+        **Note**: If sport has already a default equipment, it replaces it,
+        with the exception of the "Misc" type, in which case it is added.
+        If the sport already has 5 pieces of "Misc" equipment, an error is
+        returned.
     :<json string visibility: visibility level (``public``, ``followers_only``,
         ``private``), not mandatory (default value: ``private``)
 
@@ -383,6 +436,7 @@ def post_equipment(auth_user: User) -> Union[Tuple[Dict, int], HttpResponse]:
         - ``sport (id <sport_id>) does not exist``
         - ``invalid sport '<sport_label>' for equipment
           type '<equipment_type_label>'``
+        - ``a maximum of 5 pieces of Misc equipment can be added``
     :statuscode 401:
         - ``provide a valid auth token``
         - ``signature expired, please log in again``
@@ -472,7 +526,10 @@ def post_equipment(auth_user: User) -> Union[Tuple[Dict, int], HttpResponse]:
             },
             201,
         )
-
+    except MiscEquipmentLimitExceededException as e:
+        return MiscEquipmentInvalidPayloadErrorResponse(
+            message=e.message, sport_ids=e.sport_ids
+        )
     except (exc.IntegrityError, ValueError) as e:
         return handle_error_and_return_response(
             error=e,
@@ -553,6 +610,21 @@ def update_equipment(
         "status": "not found"
       }
 
+    - 'misc' equipment exceeding the max limit, when other equipment items
+      are already associated with sports
+
+    .. sourcecode:: http
+
+      HTTP/1.1 400 BAD REQUEST
+      Content-Type: application/json
+
+      {
+        "message": "a maximum of 5 pieces of Misc equipment can be added",
+        "sport_ids": [1],
+        "status": "limit_exceeded"
+      }
+
+
     :param string equipment_short_id: equipment short id
 
     :<json string label: a brief (less than 50 characters) label for
@@ -564,7 +636,11 @@ def update_equipment(
     :<json boolean is_active: whether or not this equipment is currently
         active (default: ``true``)
     :<json array of integers default_for_sport_ids: the default sport ids
-        to use for this equipment
+        to use for this equipment.
+        **Note**: If sport has already a default equipment, it replaces it,
+        with the exception of the "Misc" type, in which case it is added.
+        If the sport already has 5 pieces of "Misc" equipment, an error is
+        returned.
     :<json string visibility: visibility level (``public``, ``followers_only``,
         ``private``)
 
@@ -581,6 +657,7 @@ def update_equipment(
         - ``sport (id <sport_id>) does not exist``
         - ``invalid sport '<sport_label>' for equipment
           type '<equipment_type_label>'``
+        - ``a maximum of 5 pieces of Misc equipment can be added``
     :statuscode 401:
         - ``provide a valid auth token``
         - ``signature expired, please log in again``
@@ -750,7 +827,10 @@ def update_equipment(
                 "equipments": [equipment.serialize(current_user=auth_user)]
             },
         }
-
+    except MiscEquipmentLimitExceededException as e:
+        return MiscEquipmentInvalidPayloadErrorResponse(
+            message=e.message, sport_ids=e.sport_ids
+        )
     except (exc.IntegrityError, exc.OperationalError, ValueError) as e:
         return handle_error_and_return_response(
             error=e,
